Always download the signature & replace when needed
This prevents an issue where the wrong signature would be used when the
role was run for a second time (assuming that a new version of Caddy had
been released between runs) due to the original signature remaining in
place. The old signature was obviously not valid for the new version
of Caddy so caused a failure.