Support dnf5
For dnf and yum, the role includes an (empty) file with package manager
specific actions. Fedora 41 and up ship with dnf5, which ansible detects
as a seperate package manager type. In order to support dnf5, an (empty)
file named packages-dnf5.yml needs to exist.

This PR adds that file and thus support for dnf5 and Fedora 41+.